[TICKET 014] Fix code review issues
- Fix exactHitRate denominator: use finished.length instead of total so
  DNF drivers don't silently deflate the metric
- Extract Row interface to types/index.ts
- Move computeAccuracy and mapApiToRows to utils/results.ts
- Extract RaceResultsHeader and DeltaLegend into dedicated components
- ResultsPage.tsx trimmed from 273 → 133 lines (CLAUDE.md ~150 line rule)
